Versioin with more modularity, clean variable usage and flexibility

+++ functions/helpers.h
@@ -3,8 +3,8 @@
 
 #include "__init__.h"
 
-int verify_sorted(const elem_t *arr, size_t n);
-void gen_rand_num(elem_t *arr, size_t n);
-void gen_rand_num_range(elem_t *arr, size_t n, int limit);
+int verify_sorted(const elem_t *arr, long n);
+void gen_rand_num(elem_t *arr, long n);
+void gen_rand_num_range(elem_t *arr, long n, long limit);
 
 #endif
